CSC 225 Assignment Test Data Generator
Small executable which generates n random values between 1 and
1,000,000, where n is a number between 1 and 1,000,000 specified on the
command line.

If no filename is specified, output is to stdout, with values separated
by a newline, and a terminating value of -1. This makes it ideal for
piping output into Assignments 1 through 3, which deal with lists of
this format.

If a filename is given, values are output to that file separated by a
space character. This is the format expected by Assignments 1 through 3
when a filename is specified.

Usage:
testdata [n] [filename]

n: number of values to generate
filename: (Optional) name of file to output values to.